change double to int

+++ src/com/jgegroup/pacman/objects/GameObjects.java
@@ -6,7 +6,7 @@ public abstract class GameObjects {
 
     protected Position position;
     protected int radius;
-    public GameObjects(double x, double y) {
+    public GameObjects(int x, int y) {
         this.position = new Position(x, y);
     }
     public GameObjects(Position position) {
